Succeed to communicate with Navigator by using can

+++ Controller/src/f407/MDK-ARM/f407/inc/turning_module.h
@@ -18,8 +18,8 @@ typedef enum {
 
 typedef struct{
   TURNING_MODE mode;
-  float cmdAngle;
-  float shaftAngle;
+  double cmdAngle;
+  double shaftAngle;
   
   GPIO_TypeDef* zeroPort;
   uint32_t zeroMask;
